implement proper validation
